requirements.txt
colorthief==0.2.1
numpy==1.20.1
Pillow==8.1.2
